Use Mata to process replacements file
This means that Stata will not have to read the replacements file line
by line (pretty slow), and it opens the door to non-csv formats of the
replacements file that are less strictly formatted.

It also paves the way for multiple ID variables and will allow further
validation of the replacements file before replacements begin.